requirements.txt
Flask==3.0.3
imaplib2==3.6
python-dotenv==1.0.1
requests==2.32.3